pyproject.toml
[project]
name = "oz-playtype"
version = "0.1.0"
description = ""
authors = [
    {name = "kihoon",email = "nike6736@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "google-genai (>=1.59.0,<2.0.0)",
    "celery[redis] (>=5.6.2,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.12.0"
ruff = "^0.14.11"
mypy = "^1.19.1"
django-stubs = "^5.2.8"
coverage = "^7.13.1"
djangorestframework-stubs = "^3.16.6"
telnyx = "^4.10.0"

[tool.poetry]
package-mode = false

[tool.mypy]
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main"
]

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
django-redis = ">=5.4.0,<6.0.0"
redis = ">=5.0.0, <6.0.0"
telnyx = "^2.0"